Refactor src/ layout: flatten shared code, add @/* path alias
Removes confusing nesting and dead folders so the structure is
immediately readable to a new dev.

What changed:
- src/shared/components/** → src/components/ (flattened, no ui/ or form-inputs/ subfolders)
- src/shared/hooks/ → src/hooks/
- src/shared/types/database-types.ts → src/types/database-types.ts
- supabase/supabase-client.ts → src/lib/supabase.ts
- src/contexts/AuthContext.tsx → src/features/auth/AuthContext.tsx
- Deleted: src/ui/ (empty scaffolding), src/lib/hooks/ (empty),
  src/shared/ wrapper, and all unused barrel index.ts files

Added @/* → src/* path alias (tsconfig.app.json + vite.config.ts) so
cross-folder imports use @/components/Foo instead of ../../../../Foo.
All imports across the codebase updated to use the alias.

Fixed gen-types script (missing space before output redirect).

Build verified: tsc clean, 1815 modules, prerender working.
