+++ package.json
@@ -9,7 +9,7 @@
     "build": "tsc -b && vite build",
     "lint": "eslint .",
     "preview": "vite preview",
-    "gen-types": "supabase gen types typescript --project-id gbvhulvkeglzwfqccgrv> src/types/database-types.ts"
+    "gen-types": "supabase gen types typescript --project-id gbvhulvkeglzwfqccgrv > src/types/database-types.ts"
   },
   "dependencies": {
     "@supabase/supabase-js": "^2.50.3",
