+++ src/components/StepsInput.tsx
