--- src/features/auth/index.ts
@@ -1,17 +1,0 @@
-// Pages
-export { SignIn } from "./pages/SignIn";
-export { SignUp } from "./pages/SignUp";
-export { ForgotPassword } from "./pages/ForgotPassword";
-export { ResetPassword } from "./pages/ResetPassword";
-
-// Components
-export { AuthLayout } from "./components/AuthLayout";
-export { ConfirmationSent } from "./components/ConfirmationSent";
-export { GoogleSignInButton } from "./components/GoogleSignInButton";
-export { GoogleIcon } from "./components/GoogleIcon";
-
-// Hooks
-export { useGoogleAuth } from "./hooks/useGoogleAuth";
-
-// Types
-export * from "./types";
