+++ src/features/auth/pages/SignUp.tsx
@@ -1,10 +1,10 @@
 import { useState } from "react";
 import { Link, useNavigate } from "react-router";
-import { useAuth } from "../../../contexts/AuthContext";
-import { useFormValidation } from "../../../shared/hooks/useFormValidation";
-import { FormInput } from "../../../shared/components/ui/FormInput";
+import { useAuth } from "@/features/auth/AuthContext";
+import { useFormValidation } from "@/hooks/useFormValidation";
+import { FormInput } from "@/components/FormInput";
 import { GoogleSignInButton } from "../components/GoogleSignInButton";
-import { LoadingSpinner } from "../../../shared/components/ui/LoadingSpinner";
+import { LoadingSpinner } from "@/components/LoadingSpinner";
 import { AuthLayout } from "../components/AuthLayout";
 import { isSupabaseError } from "../types";
 import { ROUTES } from "../../../utils/constants";
