+++ src/features/grocery-lists/hooks/useGroceryListItems.ts
@@ -1,5 +1,5 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { supabase } from "../../../../supabase/supabase-client";
+import { supabase } from "@/lib/supabase";
 import type { GroceryListFull } from "../types";
 import { GROCERY_LIST_QUERY_KEYS } from "./query-keys";
