--- src/features/grocery-lists/index.ts
@@ -1,16 +1,0 @@
-// Pages
-export { GroceryListsPage } from "./pages/GroceryListsPage";
-export { GroceryListDetailsPage } from "./pages/GroceryListDetailsPage";
-
-// Components
-export { GroceryListDetails } from "./components/GroceryListDetails";
-export { GroceryListItems } from "./components/GroceryListItems";
-export { GroceryListList } from "./components/GroceryListList";
-export { GroceryListRecipes } from "./components/GroceryListRecipes";
-export { GroceryListTile } from "./components/GroceryListTile";
-
-// Hooks
-export * from "./hooks";
-
-// Types
-export * from "./types";
