+++ src/features/recipes/hooks/useRecipes.ts
@@ -1,6 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
-import type { Tables } from "../../../shared/types/database-types";
-import { supabase } from "../../../../supabase/supabase-client";
+import type { Tables } from "@/types/database-types";
+import { supabase } from "@/lib/supabase";
 
 export const useRecipes = () => {
   return useQuery({
