--- src/features/recipes/index.ts
@@ -1,16 +1,0 @@
-// Pages
-export { RecipesPage } from "./pages/RecipesPage";
-export { RecipeDetailsPage } from "./pages/RecipeDetailsPage";
-export { CreatePage } from "./pages/CreatePage";
-
-// Components
-export { RecipeDetails } from "./components/RecipeDetails";
-export { RecipeList } from "./components/RecipeList";
-export { RecipeTile } from "./components/RecipeTile";
-export { CreateRecipe } from "./components/CreateRecipe";
-
-// Hooks
-export * from "./hooks";
-
-// Types
-export * from "./types";
