+++ src/lib/supabase.ts
@@ -1,0 +1,7 @@
+import { createClient } from "@supabase/supabase-js";
+
+
+const supabaseUrl = import.meta.env.VITE_SUPABASE_URL
+const supabaseKey = import.meta.env.VITE_SUPABASE_ANON_KEY as string;
+
+export const supabase = createClient(supabaseUrl, supabaseKey);
