--- src/shared/components/index.ts
@@ -1,14 +1,0 @@
-// Layout components
-export { Navbar } from "./Navbar";
-export { Footer } from "./Footer";
-
-// UI components
-export { FormInput } from "./ui/FormInput";
-export { LoadingSpinner } from "./ui/LoadingSpinner";
-export { SearchInput } from "./ui/SearchInput";
-
-// Form inputs
-export { ImageUpload } from "./form-inputs/ImageUpload";
-export { IngredientMultiSelect } from "./form-inputs/IngredientMultiSelect";
-export { SearchableDropdown } from "./form-inputs/SearchableDropdown";
-export { StepsInput } from "./form-inputs/StepsInput";
