--- src/shared/hooks/forms-index.ts
@@ -1,1 +1,0 @@
-export * from "./useFormValidation"
