--- src/shared/hooks/index.ts
@@ -1,3 +1,0 @@
-export { useIngredients } from "./useIngredients";
-export { useUnits } from "./useUnits";
-export { useFormValidation } from "./useFormValidation";
