--- src/shared/index.ts
@@ -1,8 +1,0 @@
-// Components
-export * from "./components";
-
-// Hooks
-export * from "./hooks";
-
-// Types
-export * from "./types/database-types";
